Currency & holding

+++ src/main/java/com/gino/siscripto/dto/CreateWalletDTO.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class CreateWalletDTO {
     private String dni;
-    private Float saldo;
+    private BigDecimal saldo;
 }
